Create README.ko.md
Hi. Official I am a college student in Korea. I've translated the README.md file to make it easier for Korean contributors to access and contribute to the docsify.
I'm still a beginner, so I have a lot. Please let me know if there is anything you need to fix. If you wish, we will translate other documents.