rename prettierrc to .cjs and update to prettier 2
I tried to stay on prettier v1, but it doesn't work in a package with
type:module (ESM) because ESLint for some reason would then complain
about almost 4000 errors which would require a *ton* of format changes
to fix. As a result I had to update to Prettier v2 for it to work
properly, and this required only a few format updates to the codebase.